+++ packages/react-internal/src/components/Fabric/Fabric.tsx
@@ -1,9 +1,10 @@
+import * as React from 'react';
 import { styled } from '../../Utilities';
 import { FabricBase } from './Fabric.base';
 import { getStyles } from './Fabric.styles';
 import { IFabricProps, IFabricStyleProps, IFabricStyles } from './Fabric.types';
 
-export const Fabric = styled<IFabricProps, IFabricStyleProps, IFabricStyles, FabricBase>(
+export const Fabric: React.FunctionComponent<IFabricProps> = styled<IFabricProps, IFabricStyleProps, IFabricStyles>(
   FabricBase,
   getStyles,
   undefined,
