--- packages/react-next/src/components/Fabric/index.ts
@@ -1,3 +1,0 @@
-export * from './Fabric';
-export * from './Fabric.base';
-export * from './Fabric.types';
